Restructuring of make_dists_and_tree and addition of accessory methods
I rewrote make_dists_and_tree so that it takes a file path and list of host names and checks if the file at the given path is a distance matrix, sequence alignment, or tree and processes them accordingly, generating a distance matrix and/or tree if necessary. It then trims both of these according to the list of sample names and returns them to main(). Note: we discussed integrating the contents of this method into main in a future revision, as its seperation serves no significant purpose.